pyproject.toml
[tool.poetry]
name = "mini-compose"
version = "0.1.0"
description = ""
authors = ["Kalle Lämsä <1397804+coocos@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
docker = "^6.0.0"
click = "^8.1.3"
PyYAML = "^6.0"
pydantic = "^1.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"